fixed css, separated files from server/browser, modeled index route
