+++ src/browser/css/core/reset.styl
@@ -1,0 +1,2 @@
+button, html [type='button'], [type='reset'], [type='submit']
+	appearance none
