+++ src/browser/js/Root.jsx
@@ -4,8 +4,8 @@ import { Provider } from 'react-redux';
 import { syncHistoryWithStore } from 'react-router-redux';
 import { MuiThemeProvider, getMuiTheme } from 'material-ui/styles';
 
-import store from '../redux/clientStore';
-import routes from '../Router';
+import store from './redux/clientStore';
+import routes from './../../web/Router';
 
 const history = syncHistoryWithStore(browserHistory, store);
 const muiTheme = getMuiTheme();
