+++ src/browser/js/app.jsx
@@ -6,6 +6,8 @@ import immutableDevtools from 'immutable-devtools';
 
 import injectTapEventPlugin from 'react-tap-event-plugin';
 
+import Root from './Root';
+
 // immutable: logging to the console and debugging
 immutableDevtools(immutable);
 
@@ -13,8 +15,6 @@ immutableDevtools(immutable);
 // Check this repo:
 // https://github.com/zilverline/react-tap-event-plugin
 injectTapEventPlugin();
-
-import Root from './modules/Root';
 
 const view = document.getElementById('react-view');
 
@@ -29,7 +29,7 @@ render(
 
 /* eslint-disable no-undef */
 if (module.hot) {
-  module.hot.accept('./modules/Root', () => {
+  module.hot.accept('./Root', () => {
     render(
       <Root />,
       view
